+++ site-cookbooks/nginx/CHANGELOG.md
@@ -1,11 +1,11 @@
-mysql CHANGELOG
+nginx CHANGELOG
 ===============
 
-This file is used to list changes made in each version of the mysql cookbook.
+This file is used to list changes made in each version of the nginx cookbook.
 
 0.1.0
 -----
-- [your_name] - Initial release of mysql
+- [your_name] - Initial release of nginx
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.